pyproject.toml
[project]
name = "iran-persian-iptv"
version = "0.1.0"
description = "Aggregate, validate, and categorize Iranian/Persian IPTV channels into a single M3U playlist."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["ruff>=0.6"]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
